backend/requirements.txt
fastapi[standard]>=0.115.5
fastapi[test]
pydantic>=2.10
pyjwt
passlib[bcrypt]
psycopg2
alembic
sqlmodel
pyjwt[crypto]
alembic
black
pytest
Faker
